refactor(becnh_utils package): Объеденить вспомогательные модули в пакет
Объеденить вспомогательные модули в пакет
